agrego validacion de permisos a controladores y vista de boton eliminar

See merge request proyecto2021/proyectos/grupo18!12

+++ app/resources/user.py
@@ -5,14 +5,13 @@ from sqlalchemy.sql.expression import false
 from app.models.user import Rol, User, users_roles,Rol
 from app.models.elementos import Elementos
 from app.models.ordenacion import Ordenacion
-from app.helpers.auth import authenticated, check_permission
+from app.helpers.auth import assert_permission, authenticated, check_permission
 from app.db import db
 from werkzeug.security import check_password_hash, generate_password_hash
 
 # Protected resources
 def index():
-    if not authenticated(session):
-        abort(401)
+    assert_permission(session,'user_index')
 
     orden = Ordenacion.query.first()
     elem = Elementos.query.first()
@@ -21,11 +20,9 @@ def index():
     else:
         per_page = 2
     page  = int(request.args.get('page', 1,type=int))
-    if request.method == "GET":
+    
 
-        user = db.session.query(User).filter(User.email==session['user'])
-        if not check_permission(user[0].id, 'user_index'):
-            abort(401) 
+    if request.method == "GET":
 
         #aca defino por default 2 crioterios de ordenacion, por mail o descripcion
         if orden.id_orden == 1:
@@ -43,9 +40,7 @@ def index():
         return render_template("user/index.html", users=users_con_nombre)
 
 def bloqueados():
-    user = db.session.query(User).filter(User.email==session['user'])
-    if not check_permission(user[0].id, 'user_index'):
-        abort(401) 
+    assert_permission(session,'user_index') 
 
     orden = Ordenacion.query.first()
     elem = Elementos.query.first()
@@ -62,9 +57,7 @@ def bloqueados():
     return render_template("user/index.html", users=users)
 
 def no_bloqueados():
-    user = db.session.query(User).filter(User.email==session['user'])
-    if not check_permission(user[0].id, 'user_index'):
-        abort(401) 
+    assert_permission(session,'user_index')
 
     orden = Ordenacion.query.first()
     elem = Elementos.query.first()
@@ -82,9 +75,9 @@ def no_bloqueados():
 
 def edit(user_id):
     """ Edicion de usuarios """
-    if not authenticated(session):
-        abort(401)
+    assert_permission(session,'user_index')
     user = db.session.query(User).get(user_id)
+
     if request.method == 'POST':
         params = request.form   
         error = None
@@ -157,8 +150,7 @@ def edit(user_id):
 
 def create():
     """ Creacion de usuarios """
-    if not authenticated(session):
-        abort(401)
+    assert_permission(session,'user_create')
     if request.method == "GET":
         roles= db.session.query(Rol).all()
         return render_template("user/new.html", roles= roles)
@@ -202,6 +194,7 @@ def create():
             return redirect(url_for("user_create", roles = roles))
 
 def delete(user_id):
+    assert_permission(session,'user_index')
     user = db.session.query(User).get(user_id)
     db.session.delete(user)
     db.session.commit()
